Remove unnecessary Casts

+++ CSL-Traffic/AI/Vehicles/CustomAmbulanceAI.cs
@@ -84,7 +84,7 @@ namespace CSL_Traffic
             if (data.m_sourceBuilding != 0)
             {
                 BuildingManager instance = Singleton<BuildingManager>.instance;
-                if ((instance.m_buildings.m_buffer[(int)data.m_sourceBuilding].m_flags & Building.Flags.Active) == Building.Flags.None && instance.m_buildings.m_buffer[(int)data.m_sourceBuilding].m_fireIntensity == 0)
+                if ((instance.m_buildings.m_buffer[data.m_sourceBuilding].m_flags & Building.Flags.Active) == Building.Flags.None && instance.m_buildings.m_buffer[data.m_sourceBuilding].m_fireIntensity == 0)
                 {
                     return true;
                 }
